backend/requirements.txt
# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Image Processing & AI
opencv-python==4.8.1.78
pillow==10.1.0
face-recognition==1.3.0
numpy==1.25.2
scikit-image==0.22.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Email
emails==0.6.0
jinja2==3.1.2

# Utilities
python-dotenv==1.0.0
slugify==0.0.1
python-dateutil==2.8.2

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0

# File Upload & Storage
aiofiles==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Payment Processing
stripe==7.8.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1